+++ NEWS.md
@@ -4,7 +4,7 @@
 - Improved messages regarding what date/time ranges are being queried and returned by data retrieval functions.
 - `az_hourly()` now returns data from the previous hour when `start_date_time` and `end_date_time` are not supplied rather than returning the previous day of hourly data.
 - `azmet` is now much more verbose, printing messages about which data are requested and which data are returned.
-
+- Requests for data before Jan 1, 2021 now error, since these data are not on the AZMet API (yet).
 
 # azmetr 0.2.1